requirements.txt
streamlit>=1.28.0
pandas>=2.1.0 
numpy>=1.25.0
plotly==5.15.0
python-dotenv==1.0.0   